requirements.txt
praw==7.6.1
python-dotenv==0.21.0
requests==2.28.1